els-ASAP auto commit file/tag gnutls-3.3.8-14.el7_2.src.rpm

+++ src/src/certtool-common.h
@@ -111,8 +111,8 @@ const char *get_password(common_info_st * cinfo, unsigned int *flags,
 			 int confirm);
 
 extern unsigned char *lbuffer;
-extern unsigned long lbuffer_size;
+extern int lbuffer_size;
 
-void fix_lbuffer(unsigned long);
+void fix_lbuffer(unsigned);
 
 #endif
